refactor: Money 클래스 제거 및 PurchaseAmount 독립
- Money.java 삭제 (사용되지 않는 추상화)
- PurchaseAmount를 독립적인 클래스로 리팩토링
- 상속 관계 제거 (extends Money → 독립)
- protected getAmount() 제거 (완전한 캡슐화)
- validate() 메서드로 검증 로직 통합
- calculateReturnRate()를 PurchaseAmount로 이동

개선 효과:
- YAGNI 원칙 준수 (필요 없는 추상화 제거)
- 클래스 수 감소 (Money 제거)
- 상속 복잡도 제거 (단순한 구조)
- 캡슐화 강화 (private amount)
- 책임 명확화 (모든 로직이 PurchaseAmount에)

설계 철학:
"Money는 PurchaseAmount만 사용했고, 재사용 가능성이 없었다.
사용되지 않는 추상화는 복잡도만 증가시키므로 제거한다.
필요할 때 만드는 것이 미리 만드는 것보다 낫다."

Trade-off:
검증 로직이 Money에서 PurchaseAmount로 이동했지만,
Money를 어디서도 쓰지 않으므로 중복이 아니라 응집도 증가.
